Extract custom fields to components
